Fix "space in name" for retrieving imgs and allow two hovermaps on same page
Fix to work with spaces in path to image or image filename, line 13: encodeURI function.

Fix to allow multiple hover maps on the same page, which I thought was supported, but I see now why it wasn't. Line 12: Was referencing $('.mapImage'), of which is the same class for every hover map on the page. I added the $(this).siblings() function so we could target only the .mapImage that is adjacent to the currently hovered-over hotspot.